requirements.txt
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.1
rich>=13.7.0
tyro>=0.8.4
pyyaml>=6.0.1
jupyterlab>=4.1.0
ipykernel>=6.29.0
